Set loading state when data loads or error occurs to false
Instead of setting the loading state to false in a finally block which
could potentially run before the fetch request is finished and data is
set, set the loading state to false when data actually gets fetched or
an error occurs.

This was done since the loading state was getting set to false and and
hiding the loading message (which causes the site fetched data to be
displayed) before the fetch request is actually finished the product
state is set properly causing a null exception.